Moves to securitytags, adds checks.
