+++ src/core/templatetags/roles.py
@@ -1,9 +1,9 @@
 from django import template
 
 from core import models
-from core.middleware import GlobalRequestMiddleware
+
 from submission import models
-from utils import setting_handler
+
 
 register = template.Library()
 
@@ -50,30 +50,3 @@ def role_id(request, role_slug):
         return role.pk
     except models.Role.DoesNotExist:
         return 0
-
-
-@register.filter
-def se_can_see_pii(value, article):
-    # Before doing anything, check the setting is enabled:
-    se_pii_filter_enabled = setting_handler.get_setting(
-        setting_group_name='permission',
-        setting_name='se_pii_filter',
-        journal=article.journal,
-    ).processed_value
-
-    if not se_pii_filter_enabled:
-        return value
-
-    # Check if the user is an SE and return an anonymised value.
-    # If the user is not a section editor we assume they have permission
-    # to view the actual value.
-    request = GlobalRequestMiddleware.get_current_request()
-    stages = [
-        models.STAGE_UNASSIGNED,
-        models.STAGE_UNDER_REVIEW,
-        models.STAGE_UNDER_REVISION,
-    ]
-    if request.user in article.section_editors() and article.stage in stages:
-        return 'Value Anonymised'
-    else:
-        return value
