+++ src/templates/admin/review/home.html
@@ -1,6 +1,6 @@
 {% extends "admin/core/base.html" %}
 
-{% load securitytags %}
+{% load securitytags roles %}
 
 {% block title %}Articles in Review{% endblock %}
 
@@ -47,7 +47,7 @@
                             <td>{{ article.pk }}</td>
                             <td><a href="{% url 'review_in_review' article.pk %}">{{ article.safe_title }}</a></td>
                             <td>{{ article.date_submitted }}</td>
-                            <td>{{ article.correspondence_author.full_name }}</td>
+                            <td>{{ article.correspondence_author.full_name|se_can_see_pii:article }}</td>
                             <td>{% for editor in article.editors %}{{ editor.editor.full_name }}{% if not forloop.last %}, {% endif %}{% endfor %}</td>
                             <td>{{ article.section.name }}</td>
                             <td>{% if article.projected_issue %}{{ article.projected_issue.display_title }}{% else %}
