+++ src/templates/admin/review/in_review.html
@@ -1,11 +1,9 @@
 {% extends "admin/core/base.html" %}
-{% load static %}
-{% load itertools %}
-{% load roles %}
+{% load static itertools roles securitytags %}
 
 {% block title %}Review {{ article.title }}{% endblock %}
 {% block title-section %}Peer Review{% endblock %}
-{% block title-sub %}#{{ article.pk }} / {{ article.correspondence_author.last_name }} / {{ article.safe_title }}{% endblock %}
+{% block title-sub %}#{{ article.pk }} / {{ article.correspondence_author.last_name|se_can_see_pii:article }} / {{ article.safe_title }}{% endblock %}
 
 {% block breadcrumbs %}
     {{ block.super }}
