+++ src/templates/admin/review/unassigned_article.html
@@ -1,11 +1,9 @@
 {% extends "admin/core/base.html" %}
-{% load static %}
-{% load roles %}
-{% load i18n %}
+{% load static roles i18n securitytags %}
 
 {% block title %}Unassigned {{ article.title }}{% endblock %}
 {% block title-section %}Unassigned{% endblock %}
-{% block title-sub %}#{{ article.pk }} / {{ article.correspondence_author.last_name }} / {{ article.safe_title }}{% endblock %}
+{% block title-sub %}#{{ article.pk }} / {{ article.correspondence_author.last_name|se_can_see_pii:article }} / {{ article.safe_title }}{% endblock %}
 
 {% block breadcrumbs %}
     {{ block.super }}
@@ -33,7 +31,7 @@
                         <td>{{ article.section.name }}</td>
                         <td>{{ article.owner.full_name }}</td>
                         <td>{{ article.license.short_name }}</td>
-                        <td>{{ article.language }}</td>
+                        <td>{{ article.get_language_display }}</td>
                     </tr>
                     <tr>
                         <th colspan="2">Started</th>
